Add LangtonsAntExample showing coin mechanics
